feat(workflow)!: edited docker-compose & workflow
Changed docker-compose, created image with a project and is a pushed on DockerHub.
Changed language's model - altered maxlenght of the `name` field on 30 symbols.
Changed a little README.MD.
Changed config file for nginx in a part added information about of a
static files for Redoc.
Changed workflow file - added process of creating and pushing a image of the project after
merging with main branch.